migrate utilities functions of AitoClient to api
